+++ src/frontend/Orion/DataManagement/PayerProjects/PayerProject/PayerHistoricalImport/ImportSection.js
@@ -2,14 +2,21 @@ import React, { useState, useRef } from 'react'
 import XLSX from 'xlsx'
 import PropTypes from 'prop-types'
 import { useMutation } from '@apollo/react-hooks'
-
 import styled from '@emotion/styled'
+import Alert from '@material-ui/lab/Alert'
 
 import { UPLOAD_SHEET } from '../../../../../api/mutations'
+
+import Button from '../../../../../components/Button'
+import FieldLabel from '../../../../../components/FieldLabel'
+import Spinner from '../../../../../Phoenix/shared/Spinner'
 
 import Spacing from '../../../../../utils/spacing'
 import FontSpace from '../../../../../utils/fontspace'
 import Color from '../../../../../utils/color'
+import alertStatuses from '../../../../../utils/alertStatuses'
+
+const { ERROR, INFO, SUCCESS } = alertStatuses
 
 const ImportSectionWrapper = styled.div({
   display: 'flex',
@@ -24,31 +31,33 @@ const Header = styled.h1({
   ...FontSpace.FS4
 })
 
-// TODO: Replace with button component
-const Button = styled.button({
-  color: Color.WHITE,
-  backgroundColor: Color.BLUE,
-  ...FontSpace.FS2,
-  padding: `${ Spacing.S2 } ${ Spacing.S3 }`,
-  borderRadius: '4px',
-  fontWeight: 500,
-  width: 'fit-content',
-  ':hover': {
-    opacity: 0.6,
-  }
-}, ({ disabled }) => ({
-  ':hover': {
-    cursor: disabled ? 'not-allowed' : 'pointer'
-  }
-}))
-
 // TODO: Replace Input with input component
 const InputWrapper = styled.div({
   margin: `${ Spacing.S7 } 0px`,
 })
 
+const LoadingWrapper = styled.div({
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+})
+
+const alertMessageMap = {
+  success: 'Import Successful',
+  error: 'Import Failed. Please try importing again',
+  info: (
+    <LoadingWrapper>
+      <span style={{ marginRight: Spacing.S3 }}>
+        Importing
+      </span>
+      <Spinner size={14} />
+    </LoadingWrapper>
+  )
+}
+
 const ImportSection = ({
-  projectId
+  projectId,
+  setValidationErrorsAndWarnings,
 }) => {
   const fileInputRef = useRef(null)
 
@@ -56,17 +65,24 @@ const ImportSection = ({
   const [isWorkbookUploaded, setIsWorkbookUploaded] = useState(false)
   const [sheetNames, setSheetNames] = useState([])
   const [timestamp, setTimestamp] = useState(null)
+  const [alertStatus, setAlertStatus] = useState({
+    status: null, message: null
+  })
   
   const [uploadSheet] = useMutation(UPLOAD_SHEET, {
     onCompleted: ({ uploadSheet: importFeedback }) => {
-      alert(importFeedback.join('\n'))
-      // setLoading(false)
-      },
-      onError: errorMessage => {
-        alert(errorMessage)
-        // setLoading(false)
-        // setErrors(errorMessage.toString())
-      },
+      setAlertStatus({
+        status: SUCCESS,
+        message: alertMessageMap[SUCCESS]
+      })
+    },
+    onError: (errorMessage, ...rest) => {
+      setValidationErrorsAndWarnings(errorMessage.message)
+      setAlertStatus({
+        status: ERROR,
+        message: alertMessageMap[ERROR]
+      })
+    },
   })
 
   const onFileAdded = () => {
@@ -110,48 +126,57 @@ const ImportSection = ({
 
     console.log(sheetData)
 
+    setAlertStatus({
+      status: INFO,
+      message: alertMessageMap[INFO]
+    })
+
     uploadSheet({ variables: { input: sheetData } })
   }
 
   const shouldDisableButton = !isWorkbookUploaded || !timestamp
+  const buttonHoverStyle = { cursor: shouldDisableButton ? 'not-allowed' : 'pointer' }
 
   return (
     <ImportSectionWrapper>
       <Header>Import Data</Header>
       <InputWrapper>
-        <b>Pick an Excel file:</b>
-        <div>
-          <input
-            ref={fileInputRef}
-            type="file"
-            multiple
-            onChange={onFileAdded}
-          />
-        </div>
+        <FieldLabel>Pick an Excel File</FieldLabel>
+        <input
+          ref={fileInputRef}
+          type="file"
+          multiple
+          onChange={onFileAdded}
+        />
       </InputWrapper>
       <InputWrapper>
-        <label htmlFor="select-timestamp"><b>Select Timestamp</b></label>
-        <div>
-          <input
-            type="date"
-            id="select-timestamp"
-            name="select-timestamp"
-            onChange={handleDateSelection}
-          />
-        </div>
+        <FieldLabel>Select Timestamp</FieldLabel>
+        <input
+          type="date"
+          id="select-timestamp"
+          name="select-timestamp"
+          onChange={handleDateSelection}
+        />
       </InputWrapper>
       <Button
-        disabled={shouldDisableButton}
+        buttonStyle={{ width: 'fit-content', marginBottom: Spacing.S7 }}
+        hoverStyle={buttonHoverStyle}
         onClick={handleSubmit}
       >
         Import File
       </Button>
+      {alertStatus.status && (
+        <Alert severity={alertStatus.status}>
+          { alertStatus.message }
+        </Alert>
+      )}
     </ImportSectionWrapper>
   )
 }
 
 ImportSection.propTypes = {
   projectId: PropTypes.string.isRequired,
+  setValidationErrorsAndWarnings: PropTypes.func.isRequired,
 }
 
 export default ImportSection